requirements.txt
# Coinbase Algorithmic Trading Bot - Requirements

# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0

# Coinbase SDK
coinbase-advanced-py>=1.2.0

# Data analysis and technical indicators
pandas>=2.0.0
numpy>=1.24.0
pandas-ta>=0.3.14b0

# Database
# SQLite is built into Python, no extra package needed

# Optional: for enhanced features
# requests>=2.31.0
# aiohttp>=3.8.0

# Development and testing (optional)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.0.0
# flake8>=6.0.0
